+++ src/components/Navbar.jsx
@@ -137,44 +137,6 @@ const Component = styled(Box) ({
 
 const NavBar = styled(Box) ({
     display: 'flex',
-    width: '100%',
-    justifyContent: 'center',
-    // background: '#f5f5f5',
-    alignItems: 'center',
-    marginLeft: '0 auto',
-    '& > *': {
-        marginRight: '20px',
-        fontSize: '12px',
-        padding: '0 21px',
-        Opacity: 0.8,
-        // fontWeight: 600,
-        color: '#000'
-    }
-});
-
-const Header = () => {
-    return (
-        <Component>
-            <NavBar>
-                <img src={applelogo} alt="logo"  style={{width: 20}}/>
-                {
-                    navData.map(nav  => (
-                        <Typography> {nav}</Typography>
-                       
-                    ))  
-                }
-                <SearchIcon  fontSize='small'/>
-                <WorkOutlineIcon fontSize='small' />
-            </NavBar>
-        </Component>
-        )
-        // fafhvmhfg,myftkdsFDBHSBVDKJBVdqwKJGBHJKGBH,GKYHLhvmgjh,ygguYWSQWSDQsxadwfiqwddqsd    cqsfdgafwfc;io   jwqdc   qj qjeifj   ;ojqi'  ji'jfERGFBHFGEYLA efnuefhuehf jhjlvf rfjhefwio fhefnefufhlwebnlc ruf
-    }
-export default Header;
-
-   import { Box, Typography, styled } from "@mui/material";
-
-const Component = styled(Box)({
   width: '100%',
   height: 580,
   display: 'flex',
@@ -306,146 +268,4 @@ const Section = ({ data = {} }) => {
 export default Section;
 import { Box, Typography, styled } from '@mui/material';
 
-aphy component="span" style={linkStyle}>Learn more &gt;</Typography>
-        <Typography component="span" style={linkStyle}>Buy &gt;</Typography>
-      </Typography>
-    </Component>
-  );
-};
-
-export default Section;
-import { Box, Typography, styled } from '@mui/material';
-import { Box, Typography, styled } from "@mui/material";
-
-const Component = styled(Box)({
-  width: '100%',
-  height: 580,
-  display: 'flex',
-  flexDirection: 'column',
-  alignItems: 'center',
-  paddingTop: 47,
-  objectFit: 'cover',
-  backgroundRepeat: 'no-repeat',
-  backgroundSize: 'cover'
-});
-
-const Heading = styled(Typography)({
-  fontWeight: 600,
-  fontSize: 56
-});
-
-const SubHeading = styled(Typography)({
-  fontSize: 28,
-  padding: '0 0 5px 0'
-});
-
-const linkStyle = {
-  color: '#2997ff',
-  padding: '0 30px',
-  fontSize: 21,
-  cursor: 'pointer'
-};
-
-const Section = ({ data = {} }) => {
-  const {
-    imageURL = '',
-    heading = '',
-    headColor = '#000',
-    subHeading = '',
-    subHeadColor = '#333'
-  } = data;
-
-  return (
-    <Component style={{ background: `url(${imageURL}) 50% 50% no-repeat` }}>
-      <Heading style={{ color: headColor }}>{heading}</Heading>
-      <SubHeading style={{ color: subHeadColor }}>{subHeading}</SubHeading>
-      <Typography>
-        <Typography component="span" style={linkStyle}>Learn more &gt;</Typography>
-        <Typography component="span" style={linkStyle}>Buy &gt;</Typography>
-      </Typography>
-    </Component>
-  );
-};
-
-export default Section;
-import { Box, Typography, styled } from '@mui/material';
-import { Box, Typography, styled } from '@mui/material';
-
-// Styled Components
-const Component = styled(Box)({
-    width: '100%',
-    height: 580,
-    display: 'flex',
-    flexDirection: 'column',
-    alignItems: 'center',
-    paddingTop: 47,
-    objectFit: 'cover',
-    backgroundRepeat: 'no-repeat',
-    backgroundSize: 'cover'
-});
-
-const Heading = styled(Typography)({
-    fontWeight: 600,
-    fontSize: 56
-});
-
-const SubHeading = styled(Typography)({
-    fontSize: 28,
-    padding: '0 0 5px 0'
-});
-
-const linkStyle = {
-    color: '#2997ff',
-    padding: '0 30px',
-    fontSize: 21,
-    cursor: 'pointer'
-};
-
-// Functional ComponentwedqwQWD DXWQDDdqwsqdqdqwdqc   wdwdQDQWDQW,HGGF,YGYG,ILYGsaEWD
-const Product = ({ data = {} }) => {
-    const {
-        imageURL = '',
-        heading = '',
-        headColor = '#000',
-        subHeading = '',
-        subHeadColor = '#333'
-    } = data;
-
-    return (
-        <Component style={{ background: `url(${imageURL}) 50% 50% no-repeat` }}>
-            <Heading style={{ color: headColor }}>{heading}</Heading>
-            <SubHeading style={{ color: subHeadColor }}>{subHeading}</SubHeading>
-            <Typography>
-                <Typography component="span" style={linkStyle}>Learn more &gt;</Typography>
-                <Typography component="span" style={linkStyle}>Buy &gt;</Typography>
-            </Typography>
-        </Component>
-    );
-};
-// JRBGHDERBFYAGHIHFImyftkdsFDBHSBVDKJBVdqwKJGBHJKGBHmyftkdsFDBHSBVDKJBVdqwKJGBHJKGBHGREYFghghghlasdkfga<Gcyegflge,sjfbcxvhlayergrfwefg,bscghdjfvbahdsjfgeuyiwg.jkb;hu;hlh bgugh;
-
-
-const Section = ({ data = {} }) => {
-  const {
-    imageURL = '',
-    heading = '',
-    headColor = '#000',
-    subHeading = '',
-    subHeadColor = '#333'
-  } = data;
-
-  return (
-    <Component style={{ background: `url(${imageURL}) 50% 50% no-repeat` }}>
-      <Heading style={{ color: headColor }}>{heading}</Heading>
-      <SubHeading style={{ color: subHeadColor }}>{subHeading}</SubHeading>
-      <Typography>
-        <Typography component="span" style={linkStyle}>Learn more &gt;</Typography>
-        <Typography component="span" style={linkStyle}>Buy &gt;</Typography>
-      </Typography>
-    </Component>
-  );
-};
-
-export default Section;
-import { Box, Typography, styled } from '@mui/material';
